requirements.txt
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
timm==1.0.15
grad-cam==1.5.5
numpy==1.26.4
pillow==11.2.1
matplotlib==3.10.1
opencv-python==4.11.0.86






#Total Requirements:
# aiohappyeyeballs        2.6.1
# aiohttp                 3.11.18
# aiosignal               1.3.2
# asttokens               3.0.0
# attrs                   25.3.0
# certifi                 2025.4.26
# charset-normalizer      3.4.2
# colorama                0.4.6
# comm                    0.2.2
# contourpy               1.3.2
# cycler                  0.12.1
# datasets                3.5.1
# debugpy                 1.8.14
# decorator               5.2.1
# dill                    0.3.8
# einops                  0.8.1
# executing               2.2.0
# filelock                3.18.0
# fonttools               4.57.0
# frozenlist              1.6.0
# fsspec                  2025.3.0
# grad-cam                1.5.5
# huggingface-hub         0.30.2
# idna                    3.10
# ipykernel               6.29.5
# ipython                 9.2.0
# ipython_pygments_lexers 1.1.1
# jedi                    0.19.2
# Jinja2                  3.1.6
# joblib                  1.4.2
# jupyter_client          8.6.3
# jupyter_core            5.7.2
# kiwisolver              1.4.8
# MarkupSafe              3.0.2
# matplotlib              3.10.1
# matplotlib-inline       0.1.7
# mpmath                  1.3.0
# multidict               6.4.3
# multiprocess            0.70.16
# nest-asyncio            1.6.0
# networkx                3.4.2
# numpy                   1.26.4
# opencv-python           4.11.0.86
# packaging               25.0
# pandas                  2.2.3
# parso                   0.8.4
# pillow                  11.2.1
# pip                     25.1.1
# platformdirs            4.3.7
# prompt_toolkit          3.0.51
# propcache               0.3.1
# psutil                  7.0.0
# pure_eval               0.2.3
# pyarrow                 20.0.0
# Pygments                2.19.1
# pyparsing               3.2.3
# python-dateutil         2.9.0.post0
# pytz                    2025.2
# pywin32                 310
# PyYAML                  6.0.2
# pyzmq                   26.4.0
# requests                2.32.3
# safetensors             0.5.3
# scikit-learn            1.6.1
# scipy                   1.15.2
# setuptools              80.2.0
# six                     1.17.0
# stack-data              0.6.3
# sympy                   1.13.1
# threadpoolctl           3.6.0
# timm                    1.0.15
# torch                   2.5.1+cu121
# torchaudio              2.5.1+cu121
# torchcam                0.4.0
# torchvision             0.20.1+cu121
# tornado                 6.4.2
# tqdm                    4.67.1
# traitlets               5.14.3
# ttach                   0.0.3
# typing_extensions       4.13.2
# tzdata                  2025.2
# urllib3                 2.4.0
# wcwidth                 0.2.13
# xxhash                  3.5.0
# yarl                    1.20.0